requirements.txt
fastapi==0.111.1
uvicorn==0.38.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.5.0
python-pptx==0.6.21
